[SP] Change AnimSpeed vars to floats
They were supposed to be floats from the getgo, and I'm an idiot.

+++ code/qcommon/q_shared.h
@@ -1720,8 +1720,8 @@ public:
 	int forceLowerAnim;
 	int forceUpperAnimTimer;
 	int forceLowerAnimTimer;
-	int forceUpperAnimSpeed;
-	int forceLowerAnimSpeed;
+	float forceUpperAnimSpeed;
+	float forceLowerAnimSpeed;
 
 #ifndef JK2_MODE
 	//FIXME: maybe allocate all these structures (saber, force powers, vehicles)
